arm: dts: linksys: rename caiman to wrt1200ac
Linksys uses internally the codename "caiman" for a device sold under
the name of "Linksys WRT1200AC", which is already reflected in the
device tree `model`. However the `compatible` list only contains the
codename `linksys,caiman` which does not relate to the common name.

This patch renames the `dts` file to the sanitized model name and
prepends `linksys,wrt1200ac` to the `compatible` list to simplify the
device relation.

Signed-off-by: Paul Spooren <mail@aparcar.org>