arm: dts: linksys: rename rango to wrt3200acm
Linksys uses internally the codename "rango" for a device sold under the
name of "Linksys WRT3200ACM", which is already reflected in the device
tree `model`. However the `compatible` list only contains the codename
`linksys,rango` which does not relate to the common name.

This patch renames the `dts` file to the sanitized model name and
prepends `linksys,wrt3200acm` to the `compatible` list to simplify the
device relation.

Signed-off-by: Paul Spooren <mail@aparcar.org>