arm: dts: linksys: rename shelby to wrt1900acs
Linksys uses internally the codename "shelby" for a device sold under
the name of "Linksys WRT1900ACS", which is already reflected in the
device tree `model`. However the `compatible` list only contains the
codename `linksys,shelby` which does not relate to the common name.

This patch renames the `dts` file to the sanitized model name and
prepends `linksys,wrt1900acs` to the `compatible` list to simplify the
device relation.

Signed-off-by: Paul Spooren <mail@aparcar.org>