arm: dts: linksys: rename mamba to wrt1900ac
Linksys uses internally the codename "mamba" for a device sold under
the name of "Linksys WRT1900AC", which is already reflected in the
device tree `model`. However the `compatible` list only contains the
codename `linksys,mamba` which does not relate to the common name.

This patch renames the `dts` file to the sanitized model name and
prepends `linksys,wrt1900ac` to the `compatible` list to simplify the
device relation.

Signed-off-by: Paul Spooren <mail@aparcar.org>