arm: dts: linksys: rename cobra to wrt1900ac-v2
Linksys uses internally the codename "cobra" for a device sold under the
name of "Linksys WRT1900ACv2", which is already reflected in the device
tree `model`. However the `compatible` list only contains the codename
`linksys,cobra` which does not relate to the common name.

This patch renames the `dts` file to the sanitized model name and
prepends `linksys,wrt1900ac-v2` to the `compatible` list to simplify the
device relation.

For consistency with other models containing a version number a space
was added in the DT `model` to sepparate the *v2* and also reflected in
the `dts` filename containing `-v2`.

Signed-off-by: Paul Spooren <mail@aparcar.org>